Reverted cs_safe back to previous behavior. Added jax and jaxlib as docs dependencies. Added documentation on the use of openmdao.math within components.
